requirements.txt
alembic==0.6.2
blinker==1.3
dicttoxml==1.6.6
Flask==0.10.1
Flask-Bootstrap==3.0.3.1
Flask-HTTPAuth==3.2.4
Flask-Login==0.2.7
Flask-Mail==0.9.0
Flask-Migrate==1.1.0
Flask-Moment==0.2.1
Flask-Script==0.6.6
Flask-SQLAlchemy==2.1
Flask-WTF==0.9.4
itsdangerous==0.23
Jinja2==2.7.1
Mako==0.9.1
MarkupSafe==0.18
numpy==1.15.2
Pillow==5.3.0
PyMySQL==0.6.6
PyYAML==3.11
sense-hat==2.2.0
SQLAlchemy==1.1.5
validate-email==1.3
Werkzeug==0.10.4
WTForms==1.0.5
xmltodict==0.9.2
paho-mqtt==1.4.0